[REFACTOR] TestConfig 중복 제거 (#75) - global.config로 이동 및 중복 제거
